test/requirements.txt
# Test Requirements
# 테스트 실행에 필요한 Python 패키지

# 필수 라이브러리
requests>=2.31.0          # HTTP API 호출 (모든 테스트에서 사용)
certifi>=2023.7.22        # requests 의존성: SSL 인증서
charset-normalizer>=3.2.0  # requests 의존성: 문자 인코딩 감지
idna>=3.4                  # requests 의존성: Internationalized Domain Names
urllib3>=2.0.4             # requests 의존성: HTTP 클라이언트 라이브러리
pika>=1.3.2                # RabbitMQ 클라이언트 (큐 조회, DLQ 확인용)
Pillow>=10.0.0             # 이미지 생성 및 처리 (테스트 이미지 생성용)

# 선택 라이브러리
sentence-transformers>=2.2.0  # 텍스트 유사도 계산 (test_03_multi_candidate_test.py)
numpy>=1.24.0                  # sentence-transformers 의존성

# ==========================================
# 설치 방법
# pip install -r requirements.txt


